More React and more routing
