+++ Assignment4/Assignment4Final/ClientApp/public/index.html
@@ -20,7 +20,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>Assignment4Final</title>
+    <title>Assignment4-The ErrorList</title>
   </head>
   <body>
     <noscript>
