+++ Assignment4/Assignment4Final/ClientApp/src/components/Questions/QuestionEdit.js
@@ -1,0 +1,19 @@
+import { Form, Button, Col, Row, FloatingLabel, Stack , Table  } from 'react-bootstrap';
+
+import {React,useState,useEffect} from 'react';
+
+import axios from 'axios';
+import { Link } from 'react-router-dom';
+
+
+
+ function Edit()
+{
+
+
+
+}
+
+
+
+export default Edit;
